source/IoTEdge/EdgeInspect/modules/MQTTBridgeModule/requirements.txt
azure-iot-device~=2.7.0
paho-mqtt
